requirements.txt
QtAwesome>=0.4.4
tensorly>=0.2.0
numpy>=1.13.1
scikit_image>=0.13.0
matplotlib>=2.0.2
seaborn>=0.8
pycuda>=2017.1.1
requests>=2.14.2
dask>=0.15.2
h5py>=2.7.0
pytest>=3.2.1
Cython>=0.27.2
PyQt5>=5.9
hug>=2.3.2
mrcfile>=1.0.1
parse>=1.8.2
scikit_learn>=0.19.1
wget>=3.2
PyYAML>=3.12
begins>=0.9
vispy>=0.5.1

toolz>=0.8.2
